Created new component InfoDialog that used to be a v-dialog within the Audit.
